+++ tests/ImageSharp.Benchmarks/Color/Bulk/ToRgba32Bytes.cs
@@ -1,16 +1,18 @@
-// Copyright (c) Six Labors and contributors.
+﻿// Copyright (c) Six Labors and contributors.
 // Licensed under the Apache License, Version 2.0.
 
-// ReSharper disable InconsistentNaming
-
+using System;
 using System.Buffers;
 using BenchmarkDotNet.Attributes;
+
 using SixLabors.ImageSharp.Memory;
 using SixLabors.ImageSharp.PixelFormats;
 
+// ReSharper disable InconsistentNaming
+
 namespace SixLabors.ImageSharp.Benchmarks.ColorSpaces.Bulk
 {
-    public abstract class ToXyz<TPixel>
+    public abstract class ToRgba32Bytes<TPixel>
         where TPixel : struct, IPixel<TPixel>
     {
         private IMemoryOwner<TPixel> source;
@@ -27,7 +29,7 @@ namespace SixLabors.ImageSharp.Benchmarks.ColorSpaces.Bulk
         {
             this.configuration = Configuration.Default;
             this.source = this.configuration.MemoryAllocator.Allocate<TPixel>(this.Count);
-            this.destination = this.configuration.MemoryAllocator.Allocate<byte>(this.Count * 3);
+            this.destination = this.configuration.MemoryAllocator.Allocate<byte>(this.Count * 4);
         }
 
         [GlobalCleanup]
@@ -37,9 +39,28 @@ namespace SixLabors.ImageSharp.Benchmarks.ColorSpaces.Bulk
             this.destination.Dispose();
         }
 
+        //[Benchmark]
+        public void Naive()
+        {
+            Span<TPixel> s = this.source.GetSpan();
+            Span<byte> d = this.destination.GetSpan();
+
+            for (int i = 0; i < this.Count; i++)
+            {
+                TPixel c = s[i];
+                int i4 = i * 4;
+                Rgba32 rgba = default;
+                c.ToRgba32(ref rgba);
+                d[i4] = rgba.R;
+                d[i4 + 1] = rgba.G;
+                d[i4 + 2] = rgba.B;
+                d[i4 + 3] = rgba.A;
+            }
+        }
+
         [Benchmark(Baseline = true)]
         public void CommonBulk() =>
-            new PixelOperations<TPixel>().ToRgb24Bytes(
+            new PixelOperations<TPixel>().ToRgba32Bytes(
                 this.configuration,
                 this.source.GetSpan(),
                 this.destination.GetSpan(),
@@ -47,14 +68,22 @@ namespace SixLabors.ImageSharp.Benchmarks.ColorSpaces.Bulk
 
         [Benchmark]
         public void OptimizedBulk() =>
-            PixelOperations<TPixel>.Instance.ToRgb24Bytes(
+            PixelOperations<TPixel>.Instance.ToRgba32Bytes(
                 this.configuration,
                 this.source.GetSpan(),
                 this.destination.GetSpan(),
                 this.Count);
     }
 
-    public class ToXyz_Rgba32 : ToXyz<Rgba32>
+    public class ToRgba32Bytes_FromRgba32 : ToRgba32Bytes<Rgba32>
+    {
+    }
+
+    public class ToRgba32Bytes_FromArgb32 : ToRgba32Bytes<Argb32>
+    {
+    }
+
+    public class ToRgba32Bytes_FromBgra32 : ToRgba32Bytes<Bgra32>
     {
     }
 }
